refactor: qc-core userquestion 도메인 패키지 정리
