+++ qc-application/src/main/java/com/eager/questioncloud/application/api/cart/CartItemAppender.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.application.api.cart;
 import com.eager.questioncloud.core.domain.cart.infrastructure.CartItemRepository;
 import com.eager.questioncloud.core.domain.cart.model.CartItem;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionRepository;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import lombok.RequiredArgsConstructor;
