+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/FeedLibraryController.java
@@ -2,7 +2,7 @@ package com.eager.questioncloud.application.api.feed;
 
 import com.eager.questioncloud.application.api.common.PagingResponse;
 import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionDetail;
+import com.eager.questioncloud.core.domain.userquestion.dto.UserQuestionDetail;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Schema;
