+++ qc-application/src/main/java/com/eager/questioncloud/application/api/feed/FeedLibraryService.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.feed;
 
 import com.eager.questioncloud.core.domain.question.common.QuestionFilter;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionDetail;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
+import com.eager.questioncloud.core.domain.userquestion.dto.UserQuestionDetail;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
