+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/QuestionPaymentGenerator.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.application.api.payment;
 import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
 import com.eager.questioncloud.core.domain.question.model.Question;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionRepository;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import java.util.List;
