+++ qc-application/src/main/java/com/eager/questioncloud/application/api/post/PostPermissionChecker.java
@@ -3,7 +3,7 @@ package com.eager.questioncloud.application.api.post;
 import com.eager.questioncloud.core.domain.post.infrastructure.PostRepository;
 import com.eager.questioncloud.core.domain.post.model.Post;
 import com.eager.questioncloud.core.domain.question.infrastructure.QuestionRepository;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionRepository;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionRepository;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import lombok.RequiredArgsConstructor;
