+++ qc-application/src/test/java/com/eager/questioncloud/application/api/hub/QuestionHubReviewRegisterTest.java
@@ -8,9 +8,9 @@ import com.eager.questioncloud.core.domain.question.model.Question;
 import com.eager.questioncloud.core.domain.question.model.QuestionContent;
 import com.eager.questioncloud.core.domain.review.infrastructure.QuestionReviewJpaRepository;
 import com.eager.questioncloud.core.domain.review.model.QuestionReview;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestion;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionEntity;
-import com.eager.questioncloud.core.domain.userquestion.UserQuestionJpaRepository;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionEntity;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionJpaRepository;
+import com.eager.questioncloud.core.domain.userquestion.model.UserQuestion;
 import java.time.LocalDateTime;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutorService;
