+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/userquestion/implement/UserQuestionUpdater.java
@@ -1,6 +1,8 @@
-package com.eager.questioncloud.core.domain.userquestion;
+package com.eager.questioncloud.core.domain.userquestion.implement;
 
 import com.eager.questioncloud.core.domain.payment.event.CompletedQuestionPaymentEvent;
+import com.eager.questioncloud.core.domain.userquestion.infrastructure.UserQuestionRepository;
+import com.eager.questioncloud.core.domain.userquestion.model.UserQuestion;
 import com.eager.questioncloud.exception.FailQuestionPaymentException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
