+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/userquestion/infrastructure/UserQuestionJpaRepository.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.userquestion;
+package com.eager.questioncloud.core.domain.userquestion.infrastructure;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
